refactor: boxes: Set completer delims for message write box.
This small commit explicitly sets delimiter for msg_write_box, in
order to support autocompletion of text with whitespaces, for both
private and stream write boxes.

Note that the delims specifed are default ones, except that whitespace
is ignored, so as to make it function properly in autocompletes.

This is a potential WIP commit, and would possibly be amended as
per the feedback.

Tests amended (comments) and added. The existing tests for
space-containing prefixes are justified with this commit.

Fixes #925 as a side effect.